renderers/metrics_exporter/requirements.txt
Flask>=2.0.0
prometheus-client>=0.17.0
psutil>=5.9.0
gitpython>=3.1.0